Move util functions to shared
