+++ tests/src/Shared/Infrastructure/PHPUnit/Comparator/AggregateRootSimilarComparator.php
@@ -8,7 +8,7 @@ use CodelyTv\Shared\Domain\Aggregate\AggregateRoot;
 use ReflectionObject;
 use SebastianBergmann\Comparator\Comparator;
 use SebastianBergmann\Comparator\ComparisonFailure;
-use function CodelyTv\Test\isSimilar;
+use function CodelyTv\Test\Shared\isSimilar;
 
 final class AggregateRootSimilarComparator extends Comparator
 {
